requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
pytest>=7.0.0